zcash_client_sqlite

Module chain

Functions for enforcing chain validity and handling chain reorgs.

Modules

init

Functions for initializing the various databases.

migrations

Structs

BlockMetaunstable

Data structure representing a row in the block metadata database.