Function zcash_client_sqlite::chain::with_blocks

pub fn with_blocks<F>(
    cache: &BlockDb, 
    from_height: BlockHeight, 
    limit: Option<u32>, 
    with_row: F
) -> Result<(), SqliteClientError> where
    F: FnMut(CompactBlock) -> Result<(), SqliteClientError>, 

Implements a traversal of limit blocks of the block cache database.

Starting at from_height, the with_row callback is invoked with each block retrieved from the backing store. If the limit value provided is None, all blocks are traversed up to the maximum height.