zcash_client_sqlite::error

Enum SqliteClientError

pub enum SqliteClientError {

    CorruptedData(String),
    Protobuf(DecodeError),
    InvalidNote,
    TableNotEmpty,
    DecodingError(ParseError),
    TransparentDerivation(Error),
    TransparentAddress(TransparentCodecError),
    DbError(Error),
    Io(Error),
    InvalidMemo(Error),
    BlockConflict(BlockHeight),
    NonSequentialBlocks,
    RequestedRewindInvalid {
        safe_rewind_height: Option<BlockHeight>,
        requested_height: BlockHeight,
    },
    AddressGeneration(AddressGenerationError),
    AccountUnknown,
    AccountCollision(AccountUuid),
    UnknownZip32Derivation,
    KeyDerivationError(AccountId),
    BadAccountData(String),
    Zip32AccountIndexOutOfRange,
    AddressNotRecognized(TransparentAddress),
    CommitmentTree(ShardTreeError<Error>),
    CacheMiss(BlockHeight),
    ChainHeightUnknown,
    UnsupportedPoolType(PoolType),
    BalanceError(BalanceError),
    NoteFilterInvalid(NoteFilter),
    ReachedGapLimit(AccountUuid, u32),
    EphemeralAddressReuse(String, TxId),

}

The primary error type for the SQLite wallet backend.

Variants

CorruptedData(String)

Decoding of a stored value from its serialized form has failed.

Protobuf(DecodeError)

An error occurred decoding a protobuf message.

InvalidNote

The rcm value for a note cannot be decoded to a valid JubJub point.

TableNotEmpty

Illegal attempt to reinitialize an already-initialized wallet database.

DecodingError(ParseError)

A Zcash key or address decoding error

TransparentDerivation(Error)

Available on crate feature transparent-inputs only.

An error produced in legacy transparent address derivation

TransparentAddress(TransparentCodecError)

Available on crate feature transparent-inputs only.

An error encountered in decoding a transparent address from its serialized form.

DbError(Error)

Wrapper for rusqlite errors.

Io(Error)

Wrapper for errors from the IO subsystem

InvalidMemo(Error)

A received memo cannot be interpreted as a UTF-8 string.

BlockConflict(BlockHeight)

An attempt to update block data would overwrite the current hash for a block with a different hash. This indicates that a required rewind was not performed.

NonSequentialBlocks

A range of blocks provided to the database as a unit was non-sequential

RequestedRewindInvalid

A requested rewind would violate invariants of the storage layer. The payload returned with this error is (safe rewind height, requested height). If no safe rewind height can be determined, the safe rewind height member will be None.

Fields

safe_rewind_height: Option<BlockHeight>

requested_height: BlockHeight

AddressGeneration(AddressGenerationError)

An error occurred in generating a Zcash address.

AccountUnknown

The account for which information was requested does not belong to the wallet.

AccountCollision(AccountUuid)

The account being added collides with an existing account in the wallet with the given ID. The collision can be on the seed and ZIP-32 account index, or a shared FVK component.

UnknownZip32Derivation

The account was imported, and ZIP-32 derivation information is not known for it.

KeyDerivationError(AccountId)

An error occurred deriving a spending key from a seed and a ZIP-32 account index.

BadAccountData(String)

An error occurred while processing an account due to a failure in deriving the accountโs keys.

Zip32AccountIndexOutOfRange

A caller attempted to construct a new account with an invalid ZIP 32 account identifier.

AddressNotRecognized(TransparentAddress)

Available on crate feature transparent-inputs only.

The address associated with a record being inserted was not recognized as belonging to the wallet.

CommitmentTree(ShardTreeError<Error>)

An error occurred in inserting data into or accessing data from one of the walletโs note commitment trees.

CacheMiss(BlockHeight)

The block at the specified height was not available from the block cache.

ChainHeightUnknown

The height of the chain was not available; a call to WalletWrite::update_chain_tip is required before the requested operation can succeed.

UnsupportedPoolType(PoolType)

Unsupported pool type

BalanceError(BalanceError)

An error occurred in computing wallet balance

NoteFilterInvalid(NoteFilter)

A note selection query contained an invalid constant or was otherwise not supported.

ReachedGapLimit(AccountUuid, u32)

Available on crate feature transparent-inputs only.

The proposal cannot be constructed until transactions with previously reserved ephemeral address outputs have been mined. The parameters are the account UUID and the index that could not safely be reserved.

EphemeralAddressReuse(String, TxId)

Available on crate feature transparent-inputs only.

An ephemeral address would be reused. The parameters are the address in string form, and the txid of the earliest transaction in which it is known to have been used.