Struct zcash_client_sqlite::DataConnStmtCache

pub struct DataConnStmtCache<'a, P> { /* private fields */ }

The primary type used to implement WalletWrite for the SQLite database.

A data structure that stores the SQLite prepared statements that are required for the implementation of WalletWrite against the backing store.

Implementations

impl<'a, P> DataConnStmtCache<'a, P>

pub fn stmt_insert_block(
&mut self,
block_height: BlockHeight,
block_hash: BlockHash,
block_time: u32,
commitment_tree: &CommitmentTree<Node>
) -> Result<(), SqliteClientError>

Inserts information about a scanned block into the database.

impl<'a, P> DataConnStmtCache<'a, P>

pub fn stmt_update_expired(
&mut self,
height: BlockHeight
) -> Result<(), SqliteClientError>

Marks notes that have not been mined in transactions as expired, up to the given block height.