Struct zcash_client_sqlite::WalletDb

pub struct WalletDb<P> { /* private fields */ }

A wrapper for the SQLite connection to the wallet database.

Implementations

impl<P: Parameters> WalletDb<P>

pub fn for_path<F: AsRef<Path>>(path: F, params: P) -> Result<Self, Error>

Construct a connection to the wallet database stored at the specified path.

pub fn get_update_ops(
&self
) -> Result<DataConnStmtCache<'_, P>, SqliteClientError>

Given a wallet database connection, obtain a handle for the write operations for that database. This operation may eagerly initialize and cache sqlite prepared statements that are used in write operations.