zcash_client_sqlite::util

Trait Clock

pub trait Clock {
    // Required method
    fn now(&self) -> SystemTime;
}

A trait that represents the capability to read the system time.

Using implementations of this trait instead of accessing the system clock directly allows mocking with a controlled clock for testing purposes.

Required Methods

fn now(&self) -> SystemTime

Returns the current system time, according to this clock.