Module zcash_client_sqlite::wallet

Functions for querying information in the wdb database.

These functions should generally not be used directly; instead, their functionality is available via the WalletRead and WalletWrite traits.

Modules

init

Functions for initializing the various databases.

transact

Functions for creating transactions.

Traits

ShieldedOutputDeprecated

This trait provides a generalization over shielded output representations.

Functions

block_height_extremaDeprecated

Returns the minimum and maximum heights for blocks stored in the wallet database.

get_addressDeprecated

Returns the address for the account.

get_balanceDeprecated

Returns the balance for the account, including all mined unspent notes that we know about.

get_balance_atDeprecated

Returns the verified balance for the account at the specified height, This may be used to obtain a balance that ignores notes that have been received so recently that they are not yet deemed spendable.

get_block_hashDeprecated

Returns the block hash for the block at the specified height, if any.

get_commitment_treeDeprecated

Returns the commitment tree for the block at the specified height, if any.

get_nullifiersDeprecated

Retrieve the nullifiers for notes that the wallet is tracking that have not yet been confirmed as a consequence of the spending transaction being included in a block.

get_received_memoDeprecated

Returns the memo for a received note.

get_rewind_heightDeprecated

Gets the height to which the database must be rewound if any rewind greater than the pruning height is attempted.

get_sent_memoDeprecated

Returns the memo for a sent note.

get_tx_heightDeprecated

Returns the block height at which the specified transaction was mined, if any.

get_witnessesDeprecated

Returns the incremental witnesses for the block at the specified height, if any.

insert_blockDeprecated

Inserts information about a scanned block into the database.

insert_witnessDeprecated

Records the incremental witness for the specified note, as of the given block height.

is_valid_account_extfvkDeprecated

Checks whether the specified ExtendedFullViewingKey is valid and corresponds to the specified account.

mark_sapling_note_spentDeprecated

Marks a given nullifier as having been revealed in the construction of the specified transaction.

prune_witnessesDeprecated

Removes old incremental witnesses up to the given block height.

put_received_noteDeprecated

Records the specified shielded output as having been received.

put_tx_dataDeprecated

Inserts full transaction data into the database.

put_tx_metaDeprecated

Inserts information about a mined transaction that was observed to contain a note related to this wallet into the database.

update_expired_notesDeprecated

Marks notes that have not been mined in transactions as expired, up to the given block height.