Function zcash_client_sqlite::wallet::get_commitment_tree

pub fn get_commitment_tree<P>(
    wdb: &WalletDb<P>, 
    block_height: BlockHeight
) -> Result<Option<CommitmentTree<Node>>, SqliteClientError>

👎 Deprecated:
This function will be removed in a future release. Use zcash_client_backend::data_api::WalletRead::get_commitment_tree instead.

Returns the commitment tree for the block at the specified height, if any.

Examples

use tempfile::NamedTempFile;
use zcash_primitives::consensus::{Network, H0};
use zcash_client_sqlite::{
    WalletDb,
    wallet::get_commitment_tree,
};

let data_file = NamedTempFile::new().unwrap();
let db = WalletDb::for_path(data_file, Network::TestNetwork).unwrap();
let tree = get_commitment_tree(&db, H0);