Function zcash_client_sqlite::wallet::insert_sent_utxo

pub fn insert_sent_utxo<'a, P: Parameters>(
    stmts: &mut DataConnStmtCache<'a, P>, 
    tx_ref: i64, 
    output_index: usize, 
    account: AccountId, 
    to: &TransparentAddress, 
    value: Amount
) -> Result<(), SqliteClientError>

👎 Deprecated:
This method will be removed in a future update. Use zcash_client_backend::data_api::WalletWrite::store_sent_tx instead.

Inserts information about a sent transparent UTXO into the wallet database.

output_index is the index within transparent UTXOs of the transaction that contains the recipient output.