Function zcash_client_sqlite::wallet::put_sent_note

pub fn put_sent_note<'a, P: Parameters>(
    stmts: &mut DataConnStmtCache<'a, P>, 
    tx_ref: i64, 
    output_index: usize, 
    account: AccountId, 
    to: &PaymentAddress, 
    value: Amount, 
    memo: Option<&MemoBytes>
) -> Result<(), SqliteClientError>

👎 Deprecated:
This method will be removed in a future update. Use zcash_client_backend::data_api::WalletWrite::store_decrypted_tx instead.

Records information about a note that your wallet created.