zcash_client_sqlite::wallet

Module init

Functions for initializing the various databases.

Modules

migrations

Migrations for the zcash_client_sqlite wallet database.

Structs

WalletMigrator

A migrator that sets up the internal structure of the wallet database.

Enums

WalletMigrationError

Functions

init_wallet_db

Sets up the internal structure of the data database.