Function zcash_client_sqlite::wallet::init::init_accounts_table

pub fn init_accounts_table<P: Parameters>(
    wdb: &WalletDb<P>,
    keys: &HashMap<AccountId, UnifiedFullViewingKey>
) -> Result<(), SqliteClientError>

Initialises the data database with the given set of account [UnifiedFullViewingKey]s.

WARNING This method should be used with care, and should ordinarily be unnecessary. Prefer to use WalletWrite::create_account instead.

The [UnifiedFullViewingKey]s are stored internally and used by other APIs such as get_address, scan_cached_blocks, and create_spend_to_address. Account identifiers in keys MUST form a consecutive sequence beginning at account 0, and the [UnifiedFullViewingKey] corresponding to a given account identifier MUST be derived from the wallet’s mnemonic seed at the BIP-44 account path level as described by ZIP 316

Examples

use tempfile::NamedTempFile;
use secrecy::Secret;
use std::collections::HashMap;

use zcash_primitives::{
    consensus::{Network, Parameters},
    zip32::{AccountId, ExtendedSpendingKey}
};

use zcash_client_backend::{
    keys::{
        sapling,
        UnifiedFullViewingKey
    },
};

use zcash_client_sqlite::{
    WalletDb,
    wallet::init::{init_accounts_table, init_wallet_db}
};

let data_file = NamedTempFile::new().unwrap();
let mut db_data = WalletDb::for_path(data_file.path(), Network::TestNetwork).unwrap();
init_wallet_db(&mut db_data, Some(Secret::new(vec![]))).unwrap();

let seed = [0u8; 32]; // insecure; replace with a strong random seed
let account = AccountId::from(0);
let extsk = sapling::spending_key(&seed, Network::TestNetwork.coin_type(), account);
let dfvk = extsk.to_diversifiable_full_viewing_key();
let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk), None).unwrap();
let ufvks = HashMap::from([(account, ufvk)]);
init_accounts_table(&db_data, &ufvks).unwrap();