Function zcash_client_sqlite::wallet::init::init_wallet_db

pub fn init_wallet_db<P: Parameters + 'static>(
    wdb: &mut WalletDb<P>,
    seed: Option<SecretVec<u8>>
) -> Result<(), MigratorError<WalletMigrationError>>

Sets up the internal structure of the data database.

This procedure will automatically perform migration operations to update the wallet database to the database structure required by the current version of this library, and should be invoked at least once any time a client program upgrades to a new version of this library. The operation of this procedure is idempotent, so it is safe (though not required) to invoke this operation every time the wallet is opened.

It is safe to use a wallet database previously created without the ability to create transparent spends with a build that enables transparent spends (via use of the transparent-inputs feature flag.) The reverse is unsafe, as wallet balance calculations would ignore the transparent UTXOs already controlled by the wallet.

Examples

use secrecy::Secret;
use tempfile::NamedTempFile;
use zcash_primitives::consensus::Network;
use zcash_client_sqlite::{
    WalletDb,
    wallet::init::init_wallet_db,
};

let data_file = NamedTempFile::new().unwrap();
let mut db = WalletDb::for_path(data_file.path(), Network::TestNetwork).unwrap();
init_wallet_db(&mut db, Some(Secret::new(vec![]))).unwrap();