zcash_client_sqlite::wallet::init

Module migrations

Migrations for the zcash_client_sqlite wallet database.

The constants in this module cover all states of the migration DAG that have been exposed in a public crate release, in the order that crate users would have encountered them.

Omitted versions had the same migration state as the first prior version that is included.

Constants

CURRENT_LEAF_MIGRATIONS

Leaf migrations as of the current repository state.

V_0_4_0

Leaf migrations in the 0.4.0 release.

V_0_6_0

Leaf migrations in the 0.6.0 release.

V_0_8_0

Leaf migrations in the 0.8.0 release.

V_0_9_0

Leaf migrations in the 0.9.0 release.

V_0_10_0

Leaf migrations in the 0.10.0 release.

V_0_10_3

Leaf migrations in the 0.10.3 release.

V_0_11_0

Leaf migrations in the 0.11.0 release.

V_0_11_1

Leaf migrations in the 0.11.1 release.

V_0_11_2

Leaf migrations in the 0.11.2 release.

V_0_12_0

Leaf migrations in the 0.12.0 release.

V_0_13_0

Leaf migrations in the 0.13.0 release.

V_0_14_0

Leaf migrations in the 0.14.0 release.

V_0_15_0

Leaf migrations in the 0.15.0 release.

V_0_16_0

Leaf migrations in the 0.16.0 release.

V_0_16_2

Leaf migrations in the 0.16.2 release.

V_0_16_4

Leaf migrations in the 0.16.4 release.

V_0_17_0

Leaf migrations in the 0.17.0 release.

V_0_17_2

Leaf migrations in the 0.17.2 release.

V_0_17_3

Leaf migrations in the 0.17.3 release.

V_0_18_0

Leaf migrations in the 0.18.0 release.

V_0_18_5

Leaf migrations in the 0.18.5 release.

V_0_19_0

Leaf migrations in the 0.19.0 release.