Trait zcash_client_sqlite::wallet::ShieldedOutput

pub trait ShieldedOutput {
    fn index(&self) -> usize;
    fn account(&self) -> AccountId;
    fn to(&self) -> &PaymentAddress;
    fn note(&self) -> &Note;
    fn memo(&self) -> Option<&MemoBytes>;
    fn is_change(&self) -> Option<bool>;
    fn nullifier(&self) -> Option<Nullifier>;
}

👎Deprecated: This trait will be removed in a future release.

This trait provides a generalization over shielded output representations.

Required Methods

fn index(&self) -> usize

👎Deprecated: This trait will be removed in a future release.

fn account(&self) -> AccountId

👎Deprecated: This trait will be removed in a future release.

fn to(&self) -> &PaymentAddress

👎Deprecated: This trait will be removed in a future release.

fn note(&self) -> &Note

👎Deprecated: This trait will be removed in a future release.

fn memo(&self) -> Option<&MemoBytes>

👎Deprecated: This trait will be removed in a future release.

fn is_change(&self) -> Option<bool>

👎Deprecated: This trait will be removed in a future release.

fn nullifier(&self) -> Option<Nullifier>

👎Deprecated: This trait will be removed in a future release.