zcash_client_sqlite

Struct WalletDb

pub struct WalletDb<C, P> { /* private fields */ }

A wrapper for the SQLite connection to the wallet database.

Implementations

impl<P: Parameters + Clone> WalletDb<Connection, P>

pub fn for_path<F: AsRef<Path>>(path: F, params: P) -> Result<Self, Error>

Construct a connection to the wallet database stored at the specified path.

impl<C: Borrow<Connection>, P: Parameters + Clone> WalletDb<C, P>

pub fn from_connection(conn: C, params: P) -> Self

Constructs a new wrapper around the given connection.

This is provided for use cases such as connection pooling, where conn may be an &mut rusqlite::Connection.

The caller must ensure that [rusqlite::vtab::array::load_module] has been called on the connection.

impl<C: BorrowMut<Connection>, P: Parameters + Clone> WalletDb<C, P>

pub fn transactionally<F, A, E: From<Error>>(&mut self, f: F) -> Result<A, E>
where F: FnOnce(&mut WalletDb<SqlTransaction<'_>, P>) -> Result<A, E>,