Function zcash_client_sqlite::wallet::init::init_wallet_db

pub fn init_wallet_db<P>(wdb: &WalletDb<P>) -> Result<(), Error>

Sets up the internal structure of the data database.

It is safe to use a wallet database created without the ability to create transparent spends with a build that enables transparent spends via use of the transparent-inputs feature flag. The reverse is unsafe, as wallet balance calculations would ignore the transparent UTXOs controlled by the wallet. Note that this currently applies only to wallet databases created with the same version of the wallet software; database migration operations currently must be manually performed to update the structure of the database when changing versions. Integrated migration utilities will be provided by a future version of this library.

Examples

use tempfile::NamedTempFile;
use zcash_primitives::consensus::Network;
use zcash_client_sqlite::{
    WalletDb,
    wallet::init::init_wallet_db,
};

let data_file = NamedTempFile::new().unwrap();
let db = WalletDb::for_path(data_file.path(), Network::TestNetwork).unwrap();
init_wallet_db(&db).unwrap();