zcash_client_sqlite::wallet::init

Function init_wallet_db

pub fn init_wallet_db<C: BorrowMut<Connection>, P: Parameters + 'static>(
    wdb: &mut WalletDb<C, P>,
    seed: Option<SecretVec<u8>>,
) -> Result<(), MigratorError<Uuid, WalletMigrationError>>

Sets up the internal structure of the data database.

This procedure will automatically perform migration operations to update the wallet database to the database structure required by the current version of this library, and should be invoked at least once any time a client program upgrades to a new version of this library. The operation of this procedure is idempotent, so it is safe (though not required) to invoke this operation every time the wallet is opened.

In order to correctly apply migrations to accounts derived from a seed, sometimes the optional seed argument is required. This function should first be invoked with seed set to None; if a pending migration requires the seed, the function returns Err(schemerz::MigratorError::Migration { error: WalletMigrationError::SeedRequired, .. }). The caller can then re-call this function with the necessary seed.

Note that currently only one seed can be provided; as such, wallets containing accounts derived from several different seeds are unsupported, and will result in an error. Support for multi-seed wallets is being tracked in zcash/librustzcash#1284.

When the seed argument is provided, the seed is checked against the database for relevance: if any account in the wallet for which Account::source is AccountSource::Derived can be derived from the given seed, the seed is relevant to the wallet. If the given seed is not relevant, the function returns Err(schemerz::MigratorError::Migration { error: WalletMigrationError::SeedNotRelevant, .. }) or Err(schemerz::MigratorError::Adapter(WalletMigrationError::SeedNotRelevant)).

We do not check whether the seed is relevant to any imported account, because that would require brute-forcing the ZIP 32 account index space. Consequentially, imported accounts are not migrated.

It is safe to use a wallet database previously created without the ability to create transparent spends with a build that enables transparent spends (via use of the transparent-inputs feature flag.) The reverse is unsafe, as wallet balance calculations would ignore the transparent UTXOs already controlled by the wallet.

Examples

use zcash_protocol::consensus::Network;
use zcash_client_sqlite::{
    WalletDb,
    wallet::init::{WalletMigrationError, init_wallet_db},
};

let mut db = WalletDb::for_path(get_data_db_path(), Network::TestNetwork)?;
match init_wallet_db(&mut db, None) {
    Err(e)
        if matches!(
            e.source().and_then(|e| e.downcast_ref()),
            Some(&WalletMigrationError::SeedRequired)
        ) =>
    {
        let seed = load_seed()?;
        init_wallet_db(&mut db, Some(seed))
    }
    res => res,
}?;